+++ source/editor/gui/tab/camera/postprocessing/UnrealBloomPassNode.js
@@ -16,7 +16,7 @@ function UnrealBloomPassNode(parent)
 	this.add(this.strength);
 	this.nextRow();
 
-	this.addText("Radius");
+	this.addText(Locale.radius);
 	this.radius = new NumberBox(this);
 	this.radius.size.set(60, 18);
 	this.radius.setOnChange(function()