+++ source/editor/gui/tab/inspector/geometries/BoxGeometryForm.js
@@ -36,7 +36,7 @@ function BoxGeometryForm(form, object)
 	this.form.nextRow();
 
 	//Segments
-	this.form.addText("Segments");
+	this.form.addText(Locale.segments);
 	this.segmentsRow = new NumberRow(this.form);
 	this.segmentsRow.size.set(0, 18);
 
@@ -57,7 +57,7 @@ function BoxGeometryForm(form, object)
 
 	//Buffer
 	this.buffer = new CheckBox(this.form);
-	this.form.addText("Buffered");
+	this.form.addText(Locale.buffered);
 	this.buffer.size.set(18, 18);
 	this.buffer.setOnChange(updateGeometry);
 	this.form.add(this.buffer);