+++ source/editor/gui/tab/inspector/geometries/CylinderGeometryForm.js
@@ -16,7 +16,7 @@ function CylinderGeometryForm(form, object)
 	this.form.nextRow();
 
 	//Radius
-	this.form.addText("Radius");
+	this.form.addText(Locale.radius);
 	this.radiusRow = new NumberRow(this.form);
 	this.radiusRow.size.set(0, 18);
 
@@ -41,7 +41,7 @@ function CylinderGeometryForm(form, object)
 	this.form.nextRow();
 
 	//Segments
-	this.form.addText("Segments");
+	this.form.addText(Locale.segments);
 	this.segmentsRow = new NumberRow(this.form);
 	this.segmentsRow.size.set(0, 18);
 
@@ -60,7 +60,7 @@ function CylinderGeometryForm(form, object)
 
 	//Buffer
 	this.buffer = new CheckBox(this.form);
-	this.form.addText("Buffered");
+	this.form.addText(Locale.buffered);
 	this.buffer.size.set(18, 18);
 	this.buffer.setOnChange(updateGeometry);
 	this.form.add(this.buffer);