+++ source/editor/gui/tab/inspector/geometries/DodecahedronGeometryForm.js
@@ -16,7 +16,7 @@ function DodecahedronGeometryForm(form, object)
 	this.form.nextRow();
 
 	//Radius
-	this.form.addText("Radius");
+	this.form.addText(Locale.radius);
 	this.radius = new NumberBox(this.form);
 	this.radius.size.set(40, 18);
 	this.radius.setStep(0.1);
@@ -37,7 +37,7 @@ function DodecahedronGeometryForm(form, object)
 
 	//Buffer
 	this.buffer = new CheckBox(this.form);
-	this.form.addText("Buffered");
+	this.form.addText(Locale.buffered);
 	this.buffer.size.set(18, 18);
 	this.buffer.setOnChange(updateGeometry);
 	this.form.add(this.buffer);