+++ source/editor/gui/tab/scene-editor/sidebar/SideBar.js
@@ -341,6 +341,18 @@ SideBar.prototype.createObject = function()
 		Editor.addObject(model, self.editor.scene);
 	}, "Dodecahedron");
 
+	//Capsule
+	models.addOption(Global.FILE_PATH + "icons/models/capsule.png", function()
+	{
+		var geometry = new THREE.CapsuleBufferGeometry(0.5, 0.5, 1.0, 32, 1, 8, 8);
+		geometry.name = "capsule";
+		
+		var model = new Mesh(geometry, Editor.defaultMaterial);
+		model.name = "capsule";
+
+		Editor.addObject(model, self.editor.scene);
+	}, "Capsule");
+
 	//Parametric
 	/*models.addOption(Global.FILE_PATH + "icons/models/spline.png", function()
 	{